+++ lista4/1.9.c
@@ -10,12 +10,12 @@ FIM_ALGORITMO
 
 int main() {
   int num_notas;
-  float nota_atual, maior_nota = 0, segunda_nota = 0, acumulador_notas = 0;
+  double nota_atual, maior_nota = 0, segunda_nota = 0, acumulador_notas = 0;
   printf("Insira o numero de notas a serem digitadas: ");
   scanf("%d", &num_notas);
   for (int i = 1; i <= num_notas; i++) {
     printf("\nInsira a nota do aluno %d: ", i);
-    scanf("%f", &nota_atual);
+    scanf("%lf", &nota_atual);
     if (nota_atual > maior_nota) {
       segunda_nota = maior_nota;
       maior_nota = nota_atual;
@@ -24,7 +24,7 @@ int main() {
       segunda_nota = nota_atual;
     acumulador_notas += nota_atual;
   }
-  printf("As duas maiores notas sao: %.2f e %.2f\n", maior_nota, segunda_nota);
-  printf("A media da turma eh: %.2f\n", (float) acumulador_notas / num_notas);
+  printf("As duas maiores notas sao: %.2lf e %.2lf\n", maior_nota, segunda_nota);
+  printf("A media da turma eh: %.2lf\n", acumulador_notas / num_notas);
   return 0;
 }
